+++ src/Fields/TextInfo.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * This file holds an object to output a simple text without any form field.
+ *
+ * @package easy-settings-for-wordpress
+ */
+
+namespace easySettingsForWordPress\Fields;
+
+// prevent direct access.
+defined( 'ABSPATH' ) || exit;
+
+use easySettingsForWordPress\Field_Base;
+use easySettingsForWordPress\Setting;
+
+/**
+ * Object to handle the output a simple text without any form field.
+ */
+class TextInfo extends Field_Base {
+	/**
+	 * The type name.
+	 *
+	 * @var string
+	 */
+	protected string $type_name = 'TextInfo';
+
+	/**
+	 * Return the HTML-code to display this field.
+	 *
+	 * @param array $attr Attributes for this field.
+	 *
+	 * @return void
+	 */
+	public function display( array $attr ): void {
+		// bail if no attributes are set.
+		if ( empty( $attr ) ) {
+			return;
+		}
+
+		// bail if no setting object is set.
+		if ( empty( $attr['setting'] ) ) {
+			return;
+		}
+
+		// bail if field is not a Setting object.
+		if ( ! $attr['setting'] instanceof Setting ) {
+			return;
+		}
+
+		// output the value.
+		echo '<div data-depends="' . esc_attr( $this->get_depend() ) . '">';
+        echo wp_kses_post( $this->get_description() );
+        echo '</div>';
+	}
+}